code: move default sort function into component helper exports
